feat: simple cli
