+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "seed": "node seed.mjs",
+    "seed": "node --no-warnings seed.js",
     "cy:open:local": "cypress open --e2e --env environment=local",
     "cy:open:dev": "cypress open --e2e --env environment=dev",
     "cy:open:tst": "cypress open --e2e --env environment=tst",
@@ -32,7 +32,6 @@
   },
   "devDependencies": {
     "@inquirer/prompts": "^8.2.0",
-    "execa": "^9.6.1",
     "standard": "^17.1.2"
   },
   "standard": {
